# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Photo Sharing Platform API",
    "version": "1.0.0",
    "description": "A photography team uploads to an event, the lead curates a selection, and the client opens a\nPIN-protected gallery with no account.\n\n## Authentication\n\nServer-side sessions, not JWTs. `POST /auth/login` sets an httpOnly cookie; every request\nafter that carries it automatically. Revoking a session or removing a team member takes\neffect on the **next request**, which is the whole reason for the choice.\n\nNon-GET requests must echo the readable `csrf` cookie in an `X-CSRF-Token` header.\n\n## How to paginate\n\nEvery list endpoint is cursor-based. There are no page numbers and no `offset`, because rows\nshift under concurrent uploads and offset pagination would silently duplicate and skip photos.\n\nSend `limit`, read `pageInfo.nextCursor`, send it back as `cursor`, and stop when\n`hasNextPage` is false:\n\n```bash\ncursor=\"\"\nwhile :; do\n  page=$(curl -s -b cookies.txt \\\n    \"$API/events/$EVENT/photos?limit=100&sort=newest${cursor:+&cursor=$cursor}\")\n  echo \"$page\" | jq -r '.data[].filename'\n  [ \"$(echo \"$page\" | jq -r .pageInfo.hasNextPage)\" = \"true\" ] || break\n  cursor=$(echo \"$page\" | jq -r .pageInfo.nextCursor)\ndone\n```\n\nCursors are HMAC-signed and bound to the `sort` that minted them. Changing `sort` mid-walk\nreturns `400 INVALID_CURSOR` rather than a quietly wrong page."
  },
...
  "paths": {
...
    "/api/v1/public/galleries/{slug}": {
      "get": {
        "tags": [
          "Public Gallery"
        ],
        "summary": "Preview a private gallery",
        "description": "Returns display-safe metadata only. Draft and unknown slugs are indistinguishable.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 100
            },
            "required": true,
            "name": "slug",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Safe metadata for the PIN screen.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "photoCount": {
                      "type": "integer"
                    },
                    "coverThumbUrl": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "allowDownload": {
                      "type": "boolean"
                    },
                    "expiresAt": {
                      "type": [
                        "integer",
                        "null"
                      ]
                    },
                    "unlocked": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "title",
                    "photoCount",
                    "coverThumbUrl",
                    "allowDownload",
                    "expiresAt",
                    "unlocked"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Unknown, draft, or unpublished. These intentionally look identical.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "path",
                              "message"
                            ]
                          }
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "410": {
            "description": "The gallery expired.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "path",
                              "message"
                            ]
                          }
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/galleries/{slug}/unlock": {
      "post": {
        "tags": [
          "Public Gallery"
        ],
        "summary": "Unlock with the client PIN",
        "description": "Five failed attempts in 15 minutes are allowed; the sixth is rate-limited.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 100
            },
            "required": true,
            "name": "slug",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pin": {
                    "type": "string",
                    "pattern": "^\\d{6}$"
                  }
                },
                "required": [
                  "pin"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Unlocked. A gallery-specific, path-scoped cookie was set."
          },
          "401": {
            "description": "The PIN was wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "path",
                              "message"
                            ]
                          }
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Unknown, draft, or unpublished.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "path",
                              "message"
                            ]
                          }
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "410": {
            "description": "The gallery expired.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "path",
                              "message"
                            ]
                          }
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "The PIN is not six digits.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "path",
                              "message"
                            ]
                          }
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Too many attempts. Includes Retry-After.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "path",
                              "message"
                            ]
                          }
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/public/galleries/{slug}/lock": {
      "post": {
        "tags": [
          "Public Gallery"
        ],
        "summary": "Lock this gallery",
        "security": [
          {
            "galleryAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "slug",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "This gallery session ended."
          },
          "401": {
            "description": "Not unlocked.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "path",
                              "message"
                            ]
                          }
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}